% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_funs.r
\name{bed_to_granges}
\alias{bed_to_granges}
\title{BED to GRanges}
\usage{
bed_to_granges(file, header)
}
\arguments{
\item{file}{Location of your file}

\item{header}{logical; indicates whether or not bed file has a header}
}
\description{
This function loads a BED-like file and stores it as a GRanges object.
The tab-delimited file must be ordered as 'chr', 'start', 'end', 'id', 'score', 'strand'.
The minimal BED file must have the 'chr', 'start', 'end' columns.
Any columns after the strand column are ignored.
}
\details{
This is a forked version of the function written by Dave Tang
(https://github.com/davetang/bedr), with added support for bed files with headers
}
\examples{
bed_to_granges('my_bed_file.bed')
}
\keyword{BED}
\keyword{GRanges}
