% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_funs.r
\name{binaryCol}
\alias{binaryCol}
\title{Check if site in bed file; returns 0 or 1}
\usage{
binaryCol(sites, bedfile)
}
\arguments{
\item{sites}{A dataframe with 10 columns (CHR, POS, Sequence, Dummy Variables for the 6 mutation types, and read depth)}

\item{bedfile}{Path to bedfile}
}
\value{
Vector with binary indicators to be appended as column to input data.frame
}
\description{
Check if site in bed file; returns 0 or 1
}
