% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.r
\name{buildDF}
\alias{buildDF}
\title{Build list of dataframes for each model}
\usage{
buildDF(fitlist, data)
}
\arguments{
\item{fitlist}{list containing fitted values}

\item{data}{Data to be fitted}
}
\value{
list of fitted values
}
\description{
Build list of dataframes for each model
}
