% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.r
\name{corSE}
\alias{corSE}
\title{Compute standard error for correlations}
\usage{
corSE(corval, ct)
}
\arguments{
\item{corval}{correlation}

\item{ct}{count?}
}
\value{
list of fitted values
}
\description{
Compute standard error for correlations
}
