% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_funs.r
\name{gcCol}
\alias{gcCol}
\title{Get GC content in 10kb window}
\usage{
gcCol(sites, gcfile)
}
\arguments{
\item{sites}{A dataframe with 10 columns (CHR, POS, Sequence, Dummy Variables for the 6 mutation types, and read depth)}

\item{gcfile}{Path to file containing percent gc content}
}
\value{
Vector with gc content to be appended as column to input data.frame
}
\description{
Get GC content in 10kb window
}
