% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.r
\name{get_aggseq}
\alias{get_aggseq}
\title{get_aggseq}
\usage{
get_aggseq(data, mct)
}
\arguments{
\item{data}{directory containing the motif counts (assumed 1 per chromosome)}

\item{mct}{string or regular expression to match input files in bindir}
}
\value{
calculates K-mer mutation rates from input data
}
\description{
get_aggseq
Helper function for getData
}
\examples{
\dontrun{get_aggseq(sites, mct)}
}
\author{
Jed Carlson
}
