% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.r
\name{get_bins}
\alias{get_bins}
\title{get_bins}
\usage{
get_bins(bindir, pattern)
}
\arguments{
\item{bindir}{directory containing the motif counts (assumed 1 per chromosome)}

\item{pattern}{string or regular expression to match input files in bindir}
}
\value{
concatenates files into a single data frame
}
\description{
get_bins
Helper function for getData
}
\examples{
get_bins("/path/to/motif_counts", "motifs.txt")
}
\author{
Jed Carlson
}
