% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_sigs.r
\name{get_loads}
\alias{get_loads}
\title{get signature loadings}
\usage{
get_loads(widedat, nmfdat)
}
\arguments{
\item{widedat}{Output from get_nmf_input}

\item{nmfdat}{Output from nmf}
}
\value{
Signature loadings
}
\description{
get signature loadings
}
