% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.r
\name{get_mct}
\alias{get_mct}
\title{get_mct}
\usage{
get_mct(bins)
}
\arguments{
\item{pattern}{string or regular expression to match input files in bindir}
}
\value{
counts total number of motifs in bin directory
}
\description{
get_mct
Helper function for getData
}
\examples{
get_bins(bins)
}
\author{
Jed Carlson
}
