% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{getArgs}
\alias{getArgs}
\title{getArgs}
\usage{
getArgs(verbose = FALSE, defaults = NULL)
}
\arguments{
\item{verbose}{print verbage to screen}

\item{defaults}{a named list of defaults, optional}
}
\value{
a named list
}
\description{
commandArgs parsing (DEPRECATED)
return a named list of command line arguments
}
\details{
Usage:
call the R script thus
  ./myfile.R --args myarg=something
or
  R CMD BATCH --args myarg=something myfile.R

Then in R do
  myargs <- getArgs()
and myargs will be a named list
> str(myargs)
List of 2
$ file : chr "myfile.R"
$ myarg: chr "something"
}
\author{
Chris Wallace
}
