% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_build.r
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(summfile, singfile, bindir, maskfile, binw)
}
\arguments{
\item{summfile}{input file containing list of sites}

\item{singfile}{file containing sites annotated with sample ID}

\item{bindir}{directory containing the motif counts}

\item{maskfile}{Path to genome mask file (BED format)}

\item{binw}{Binwidth}
}
\value{
a list of four components: sites, bins, mct, & aggseq
}
\description{
Function gets basic summary data from input directory
}
\examples{
\dontrun{("sites.txt", "sites_id.txt", "/path/to/motif_counts")}
}
\author{
Jed Carlson
}
