% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.r
\name{getFits}
\alias{getFits}
\title{Build list of fitted values (with CHR/BIN names) for each model}
\usage{
getFits(modlist, data)
}
\arguments{
\item{modlist}{list containing model objects}

\item{data}{Data to be fitted}
}
\value{
list of fitted values
}
\description{
Build list of fitted values (with CHR/BIN names) for each model
}
