% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_funs.r
\name{imputeStart}
\alias{imputeStart}
\title{Function determines start of interval from value in previous row}
\usage{
imputeStart(ends)
}
\arguments{
\item{ends}{vector containing positions}
}
\value{
Vector of starting positions
}
\description{
Function determines start of interval from value in previous row
}
