% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{insetPlot}
\alias{insetPlot}
\title{Display plot with inset}
\usage{
insetPlot(main, inset, loc)
}
\arguments{
\item{main}{ggplot2 object to be displayed}

\item{inset}{How many adjacent bases?}

\item{loc}{Where to place inset}
}
\description{
Display plot with inset
}
