% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\docType{data}
\name{iwhPalette}
\alias{iwhPalette}
\title{A vector of colors}
\format{An object of class \code{character} of length 12.}
\usage{
iwhPalette
}
\description{
A vector of colors
}
\keyword{datasets}
