% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\docType{data}
\name{orderedcats1}
\alias{orderedcats1}
\title{Ordered list of mutation types}
\format{An object of class \code{character} of length 9.}
\usage{
orderedcats1
}
\description{
Ordered list of mutation types
}
\keyword{datasets}
