% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\docType{data}
\name{orderedcats2}
\alias{orderedcats2}
\title{Ordered list of mutation types}
\format{An object of class \code{character} of length 9.}
\usage{
orderedcats2
}
\description{
Ordered list of mutation types
}
\keyword{datasets}
