% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_sigs.r
\name{plot_ind_sigs}
\alias{plot_ind_sigs}
\title{plot signature contribution across individuals}
\usage{
plot_ind_sigs(sigdat)
}
\arguments{
\item{sigdat}{Signature loadings from get_loads}
}
\value{
ggplot2 object with signature by individual loads plot
}
\description{
plot signature contribution across individuals
}
