% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_sigs.r
\name{plot_loads}
\alias{plot_loads}
\title{plot signature loadings}
\usage{
plot_loads(sigloads)
}
\arguments{
\item{sigloads}{Signature loadings from get_loads}
}
\value{
ggplot2 object with signature loads plot
}
\description{
plot signature loadings
}
