% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_funs.r
\name{repCol}
\alias{repCol}
\title{Get replication timing rate for each site}
\usage{
repCol(sites, repfile)
}
\arguments{
\item{sites}{A dataframe with 10 columns (CHR, POS, Sequence, Dummy Variables for the 6 mutation types, and read depth)}

\item{repfile}{Path to file containing replication timing}
}
\value{
Vector with replication timing to be appended as column to input data.frame
}
\description{
Get replication timing rate for each site
}
