% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_process.r
\name{revcomp}
\alias{revcomp}
\title{Function to get reverse complement}
\usage{
revcomp(DNAstr)
}
\arguments{
\item{DNAstr}{String to get reverse complement of}
}
\value{
Reverse compliment of the string
}
\description{
Function to get reverse complement
}
\examples{
revcomp("ATCG")
}
