% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_process.r
\name{reverse_chars}
\alias{reverse_chars}
\title{Function to reverse sequence}
\usage{
reverse_chars(string)
}
\arguments{
\item{string}{Sequence to be reversed}
}
\value{
Reversed string
}
\description{
Used to correctly plot right flank in subsequence heatmaps
}
