% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{rrheat2}
\alias{rrheat2}
\title{Function for plotting K-mer heatmaps}
\usage{
rrheat2(dat, adj)
}
\arguments{
\item{dat}{Data to be plotted}

\item{adj}{How many adjacent bases?}
}
\value{
ggplot2 object with heatmap
}
\description{
Function for plotting K-mer heatmaps
}
