% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{rrheat3}
\alias{rrheat3}
\title{Plot relative differences between ERVs and AV rates}
\usage{
rrheat3(dat)
}
\arguments{
\item{dat}{Data to be plotted}
}
\value{
ggplot2 object with heatmap
}
\description{
Plot relative differences between ERVs and AV rates
}
