% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.r
\name{runMod}
\alias{runMod}
\title{Run negbin model for each formula specified in formlist}
\usage{
runMod(formlist, data)
}
\arguments{
\item{formlist}{List of formula objects}

\item{data}{Data on which you wish to run the model(s)}
}
\value{
List of model objects returned by glm.nb
}
\description{
Run negbin model for each formula specified in formlist
}
