% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaug.r
\docType{package}
\name{smaug-package}
\alias{smaug}
\alias{smaug-package}
\title{Helper library for smaug pipeline}
\description{
This package contains functions used in Carlson, et al (2018)
for the evaluation of germline mutation rate heterogeneity across the 
genome using extremely rare variants.
}
\author{
\strong{Maintainer}: Andy Beck \email{beckandy@umich.edu} [contributor]

Authors:
\itemize{
  \item Carlson Jedidiah
}

}
