% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{timefun}
\alias{timefun}
\title{timefun}
\usage{
timefun(script)
}
\arguments{
\item{script}{path to script you wish to source}
}
\description{
timefun
sources script with runtime info
}
\author{
Jed Carlson
}
