% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePackage.r
\name{usePackage}
\alias{usePackage}
\title{usePackage}
\usage{
usePackage(package)
}
\arguments{
\item{package}{package to be installed}
}
\description{
usePackage
Convenience function for installing and loading R packages from CRAN
}
\examples{
usePackage("tidyverse")
}
\author{
Jed Carlson
}
